+++ docs/changelog.rst
@@ -8,6 +8,7 @@ CHANGELOG
 **Bug fixes**
 
 * Fix elevation info not being computed when intervention is created (ref #1221)
+* Fix list of values for infrastructure and signage types (fixes #1223)
 
 0.27.1 (2010-10-13)
 -------------------